Split servermap.cpp/h off from map.cpp/h
